update params & set role for ticket api

+++ controllers/ticket/ticket.js
@@ -1,6 +1,6 @@
 const Ticket = require("../../models/Ticket/Ticket");
 const CategoryTicket = require("../../models/Ticket/CategoryTicket");
-
+const moment = require("moment");
 // Tạo ticket mới
 exports.createTicket = async (req, res) => {
   try {
@@ -39,7 +39,46 @@ exports.createTicket = async (req, res) => {
 // Xem tất cả ticket
 exports.getAllTickets = async (req, res) => {
   try {
-    const tickets = await Ticket.find();
+    const user = req.user;
+    const { search_value, form_date, to_date } = req.query;
+
+    // Khởi tạo query để tìm kiếm
+    let ticketsQuery = {};
+
+    // Điều kiện xác định quyền truy cập của người dùng
+    if (
+      user.role === "Admin" ||
+      user.role === "Manager" ||
+      user.role === "Employee" ||
+      user.role === "Collaborator"
+    ) {
+    } else {
+      ticketsQuery.createdBy = user.id;
+    }
+
+    if (search_value) {
+      ticketsQuery.$text = { $search: search_value };
+    }
+
+    // Bộ lọc theo form_date và to_date (ngày tháng)
+    if (form_date && to_date) {
+      const startDate = moment
+        .utc(form_date, "DD/MM/YYYY")
+        .startOf("day")
+        .toDate();
+      const endDate = moment.utc(to_date, "DD/MM/YYYY").endOf("day").toDate();
+
+      ticketsQuery.createdAt = {
+        $gte: startDate,
+        $lte: endDate,
+      };
+    }
+    // console.log("Date Query:", ticketsQuery.createdAt);
+
+    // Lấy tickets từ cơ sở dữ liệu theo query đã xây dựng
+    const tickets = await Ticket.find(ticketsQuery);
+
+    // Trả về kết quả
     res.status(200).json(tickets);
   } catch (error) {
     console.error("Error fetching tickets:", error);
@@ -50,14 +89,23 @@ exports.getAllTickets = async (req, res) => {
 // Xem chi tiết ticket
 exports.getTicketById = async (req, res) => {
   try {
+    const user = req.user;
     const { id } = req.params;
 
     const ticket = await Ticket.findById(id);
     if (!ticket) {
       return res.status(404).json({ error: "Ticket not found." });
     }
-
-    res.status(200).json(ticket);
+    if (
+      user.role === "Admin" ||
+      user.role === "Manager" ||
+      user.role === "Employee" ||
+      user.role === "Collaborator" ||
+      ticket.createdBy.toString() === user.id
+    ) {
+      return res.status(200).json(ticket);
+    }
+    return res.status(403).json({ error: "Access denied." });
   } catch (error) {
     console.error("Error fetching ticket:", error);
     res.status(500).json({ error: "Failed to fetch ticket." });
@@ -67,8 +115,15 @@ exports.getTicketById = async (req, res) => {
 // Cập nhật trạng thái ticket
 exports.updateTicketStatus = async (req, res) => {
   try {
+    const user = req.user;
     const { id } = req.params;
     const { status } = req.body;
+
+    if (user.role !== "Admin") {
+      return res.status(403).json({
+        error: "Access denied. Only Admin can update the ticket status.",
+      });
+    }
 
     // Kiểm tra trạng thái hợp lệ
     if (!["pending", "in-progress", "resolved"].includes(status)) {
@@ -98,8 +153,14 @@ exports.updateTicketStatus = async (req, res) => {
 // Xóa ticket
 exports.deleteTicket = async (req, res) => {
   try {
+    const user = req.user; // Lấy từ middleware xác thực
     const { id } = req.params;
 
+    if (user.role !== "Admin") {
+      return res
+        .status(403)
+        .json({ error: "Access denied. Only Admin can delete the ticket." });
+    }
     const ticket = await Ticket.findByIdAndDelete(id);
 
     if (!ticket) {
